+++ build3d.sh
@@ -32,9 +32,7 @@ pyinstaller \
     --collect-all direct \
     --hidden-import igdbpy \
     --hidden-import howlongtobeatpy \
-    --hidden-import igdbpy \
     --hidden-import requests \
-    --hidden-import steampy \
     --hidden-import threading \
     --hidden-import concurrent.futures \
     --hidden-import logging \